backend/pyproject.toml
[tool.poetry]
name = "rusty_bois"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.0"
graphene = "^2.1"
option = "^0.3.0"
gunicorn = "^19.9"
attrs = "^18.2"
flask-cors = "^3.0"
pyjwt = "^1.7"
passlib = "^1.7"

[tool.poetry.dev-dependencies]
pytest = "^4.2"
black = "18.9b0"
pytest-cov = "^2.6"
flake8 = "^3.7"
faker = "^1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
